Commit Action: Reverted xH.js to use a relative, rather than absolute path Commit Reasons: (a) using an absolute path prevents IE working on local filesystem, and likely any domain other than code.google.com, due to cross-domain scripting restrictions being tighter on HTCs. (b) Ideally, when visiting an Ubiquity 0.2 page, using Ubiquity 0.2 scripts is the more expected behaviour
I agree that this requires the files that use it to be in a certain location, relative to the library, which is suboptimal.  I would suggest that the final solution for this issue might be that in production, this file should be dynamically generated, such that it always corresponds to the correect absolute or at least root-relative path.