Remove the list of decoration classes from the XBL/HTC decorator URIs. The list is now figured out by introspecting the element (which may be identified for decoration using any number of means).
The decorator maintains a set of rules that it applies to each element that needs to be decorated. Rules have the following properties:
 * name - Used to identify the rule (not currently used)
 * match - The optional function that returns true iff the candidate element needs to be decorated by the rule (optional since if the only criteria for matching is element name, that is handled by the partitioning)
 * apply - The function that manipulates the list of class names

For performance, the rules are partitioned by element name. For rules that may be applicable across element names (for example, the presence of a "ref" attribute), a "wildcard" partition is used (we have no such rules at present).

I've added a NamespaceManager.getNamespaceURI(node) function (does what the name suggests).

Finally, a couple of other minor changes were made to decorate.js (same as those in r941, which is not yet in trunk):
1) The generateMozBindingStyle method will only generate the ? character if the 'objects' array is not empty
2) The getDecorationObjectNames method will never split an empty string (this results in a one element array instead of an empty array)