[REVIEW REQUEST] Adding initial WebKit support.
The decoration rules are applied to the document after successful completion of the loading sequence. We:
(a) select elements having QNames that have rules registered against them (one QName at a time), and
(b) apply decoration rules to each of the elements selected
Thus, the act of adding decoration rules also implicitly provides selection criteria (and no additional elements are selected or traversed at this time).

The above technique is abandoned if the user agent supports a smarter way of providing decoration support (see UX.hasDecorationSupport property).

When the above technique is used, it currently performs one pass over the document. This means repeats and other dynamic operations are not yet supported on WebKit.

Some other smaller changes were made along the way to accomodate WebKit.

Finally:
 * The UX flags from xforms/ie6-css-selectors-fixer.js have been moved to (new file) sniffer.js where they seem more appropriate.
 * The g_bIsInXHTMLMode flag has been replaced by UX.isXHTML

Tried samples on Safari (desktop and iPhone) and Chrome.