[Review Request] (I hope this svn copy works from eclipse) Please review these changes for issue-44: support for application/xhtml+xml content type.
The primary areas of work have been to address the html vs xml parsing dom models as noted in http://code.google.com/p/ubiquity-xforms/issues/detail?id=44&colspec=Iteration%20Owner%20ID%20Type%20Priority%20Summary%20Component.
 
The primary changes have been to 
--Remove any document.write API calls (at least, where not necessary)
--Centralize the modification of the Element.style property
--Centralize the modification for Element.className (add and remove)
--Centralize getElementById API call
In Firefox XML parsing, this content type is automatically set to applicaiton/xhtml+xml with *.xhtml extensions, the write API call isn't supported, the style and className properties are not supported for non-XHTML (or non-XUL) elements, and non-XHMTL(or non-XUL) elements with id attributes can not be determined with a getElementById invocation.
 
CSS classnames for non-XHTML elements are ignored unless using the selector syntax (equivalent to .className) such as *[class~="className"] is used instead.
 
The insertAdjacentHTML.js was renamed to firefoxUtils.js for the centralized methods for style, classname and getElementById.