[CODE REVIEW NEEDED] Fix for issue 116 There are 3 failed tests in issue 116, the first 2 test 4.3.8.a.xhtml 8.1.5.c.xhtml is fixed by flipping the capture parameter to false for the addEventListener (see comment)
The third test 
8.2.3.a.xhtml
is fixed by making hint aware of the model context when a ref is present.
In order to fix this bug, I have added the class "Context" to the hint's definition
and copied some code from "Group" to make "hint" able to get refresh from model.

I have considered adding the "Control" class to the Hint's definition but hesitate to
do that since using "Control" could result in creation of "pe-value" on all the hint elements
and this would make the simple design of Hint much more complicated. And I also think that if 
I modify the "Control" class to selectively create "pe-value" could overload the already packed
functionalities of "Control".