This is where I got to with xf:upload earlier, before switching my attentions to issue 268 for a bit.
I've got a problem in that my boilerplate code for displaying the new control is not working at all in Firefox (the label is just displayed inline). In Internet Explorer, the control is there, but it appears to be getting clipped by the outer pe-value element somehow, so is partially obscured.

I'm committing this to my changes folder so that other eyes can maybe point out where I've gone wrong. The boilerplate stuff for adding new controls is all quite mysterious to me, so I've no idea what I've missed and/or got wrong.