Harmonisation of earlier changes to build.xml with the broken housekeeping Buildbot builder.
When I first started working on this Buildbot/Ant stuff, I thought it was only the remote slaves that were using this build.xml. But I've now noticed that the housekeeping builder also uses it, hence my changes broke housekeeping at some point down the line.

So I've reinstated the Ant target that housekeeping uses and instead moved my modified Ant target to a different name. This is not the most optimal fix in the world, there is some duplication, but my main aim here is to get the things that I broke working again.

There have also been a couple of complementary changes made to master.cfg on the Buildbot server.